chore(hub): tier guard lives in the tier domain (#733) + cutover-fence doc note + batch-D changeset
Move assertTierComposition from with-lookup/indexing/unique-constraints.ts
to with-audit/tiers/index.ts (#733) — the tier-composition guard now lives
in the tier domain instead of an indexing file it only borrowed for a
grandfathered import specifier. Its Collection-constructor call site was
already removed by Arc 10 Task 1 (#724, the tiers+blobFields refusal was
superseded by a runtime read gate on collection.blob()), so this is a pure
relocation of dead code with no import changes needed in collection.ts or
collection-config.ts — collection.ts stays line-neutral (4547, ceiling 4549).

Add a docblock note to assertCutoverTierSafe (kernel/tier-visibility.ts)
from Task 4's review: the all-or-nothing pre-check guards ordinary
put()/delete() via the schema fence, but elevate()/demote()/putAtTier()
bypass the fence — a tier move racing the scan/rewrite window is not
excluded; cutover callers assume tier-quiescence like the documented
CAS-less caller (migrate-cek.ts).

Add .changeset/m34-batch-d.md for the batch's five issues (#730, #725,
#720, #718, #708).
